TASK: Move PHP8 test classes to own folder
Also avoids registering those classes if not on PHP8 or later.
