--- Neos.Flow/Tests/Functional/ObjectManagement/Fixtures/ClassWithConstructorProperties.php
@@ -1,39 +1,0 @@
-<?php
-namespace Neos\Flow\Tests\Functional\ObjectManagement\Fixtures;
-
-/*
- * This file is part of the Neos.Flow package.
- *
- * (c) Contributors of the Neos Project - www.neos.io
- *
- * This package is Open Source Software. For the full copyright and license
- * information, please view the LICENSE file which was distributed with this
- * source code.
- */
-
-use Neos\Flow\Annotations as Flow;
-
-if (PHP_MAJOR_VERSION >= 8) {
-    /**
-     * A class with PHP 8 constructor properties
-     * @Flow\Scope("prototype")
-     */
-    class ClassWithConstructorProperties
-    {
-        public function __construct(
-            public ?string $propertyA = 'Foo',
-            public int|null $propertyB = 25,
-            public \DateTime|null $propertyC = null
-        )
-        {
-        }
-    }
-} else {
-    /**
-     * A class with PHP 8 constructor properties
-     * @Flow\Scope("prototype")
-     */
-    class ClassWithConstructorProperties
-    {
-    }
-}
